+++ pkgs/dart_mcp_server/lib/src/mixins/pub_dev_search.dart
@@ -43,11 +43,9 @@ base mixin PubDevSupport on ToolsSupport {
     try {
       result = jsonDecode(await _client.read(searchUrl));
 
-      final packageNames =
-          dig<List>(result, ['packages'])
-              .take(_resultsLimit)
-              .map((p) => dig<String>(p, ['package']))
-              .toList();
+      final packageNames = dig<List>(result, [
+        'packages',
+      ]).take(_resultsLimit).map((p) => dig<String>(p, ['package'])).toList();
 
       if (packageNames.isEmpty) {
         return CallToolResult(
@@ -71,15 +69,17 @@ base mixin PubDevSupport on ToolsSupport {
       }
 
       // Retrieve information about all the packages in parallel.
-      final subQueryFutures =
-          packageNames
-              .map(
-                (packageName) => (
-                  versionListing: retrieve('api/packages/$packageName'),
-                  score: retrieve('api/packages/$packageName/score'),
-                ),
-              )
-              .toList();
+      final subQueryFutures = packageNames
+          .map(
+            (packageName) => (
+              versionListing: retrieve('api/packages/$packageName'),
+              score: retrieve('api/packages/$packageName/score'),
+              docIndex: retrieve(
+                'documentation/$packageName/latest/index.json',
+              ),
+            ),
+          )
+          .toList();
 
       // Aggregate the retrieved information about each package into a
       // TextContent.
@@ -88,6 +88,17 @@ base mixin PubDevSupport on ToolsSupport {
         final packageName = packageNames[i];
         final versionListing = await subQueryFutures[i].versionListing;
         final scoreResult = await subQueryFutures[i].score;
+        final libraryDocs = {
+          for (var object
+              in ((await subQueryFutures[i].docIndex) as List?)
+                      ?.cast<Map<String, Object?>>() ??
+                  <Map<String, Object?>>[])
+            if (!object.containsKey('enclosedBy'))
+              object['name'] as String: Uri.https(
+                'pub.dev',
+                'documentation/$packageName/latest/${object['href']}',
+              ).toString(),
+        };
         results.add(
           TextContent(
             text: jsonEncode({
@@ -97,12 +108,28 @@ base mixin PubDevSupport on ToolsSupport {
                   'latest',
                   'version',
                 ]),
-                'description': dig<String>(versionListing, [
+                'description': ?dig<String?>(versionListing, [
                   'latest',
                   'pubspec',
                   'description',
                 ]),
+                'homepage': ?dig<String?>(versionListing, [
+                  'latest',
+                  'pubspec',
+                  'homepage',
+                ]),
+                'repository': ?dig<String?>(versionListing, [
+                  'latest',
+                  'pubspec',
+                  'repository',
+                ]),
+                'documentation': ?dig<String?>(versionListing, [
+                  'latest',
+                  'pubspec',
+                  'documentation',
+                ]),
               },
+              if (libraryDocs.isNotEmpty) ...{'libraries': libraryDocs},
               if (scoreResult != null) ...{
                 'scores': {
                   'pubPoints': dig<int>(scoreResult, ['grantedPoints']),
@@ -112,19 +139,15 @@ base mixin PubDevSupport on ToolsSupport {
                     'downloadCount30Days',
                   ]),
                 },
-                'topics':
-                    dig<List>(
-                      scoreResult,
-                      ['tags'],
-                    ).where((t) => (t as String).startsWith('topic:')).toList(),
-                'licenses':
-                    dig<List>(scoreResult, ['tags'])
-                        .where((t) => (t as String).startsWith('license'))
-                        .toList(),
-                'publisher':
-                    dig<List>(scoreResult, ['tags'])
-                        .where((t) => (t as String).startsWith('publisher:'))
-                        .firstOrNull,
+                'topics': dig<List>(scoreResult, [
+                  'tags',
+                ]).where((t) => (t as String).startsWith('topic:')).toList(),
+                'licenses': dig<List>(scoreResult, [
+                  'tags',
+                ]).where((t) => (t as String).startsWith('license')).toList(),
+                'publisher': dig<List>(scoreResult, ['tags'])
+                    .where((t) => (t as String).startsWith('publisher:'))
+                    .firstOrNull,
               },
             }),
           ),